requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt 4.4.0
idna==2.10
importlib-metadata==4.5.0
Markdown==3.3.4
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.2.0
pip==21.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
setuptools==56.0.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wheel==0.36.2
zipp==3.4.1
PyJWT==1.7.1
